requirements.txt
Flask
Flask-SQLAlchemy
Flask-restful
matplotlib
PyJWT==1.7.1